job_profiling/requirements_1.txt
django
pymongo
mongoengine
django-mongoengine
django-rest-framework-mongoengine
djangorestframework
pandas
numpy
scikit-learn
graphviz
scikit-survival
scipy
lifelines
matplotlib
joblib